services/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg[binary]==3.1.17
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aioredis==2.0.1
redis==5.0.1
celery==5.3.6
httpx==0.26.0
python-dotenv==1.0.0
asyncpg==0.29.0
cryptography==42.0.0
tenacity==8.2.3
prometheus-client==0.19.0
slowapi==0.1.9
boto3==1.34.28
minio==7.2.3
